Split acid-state-tests package out of acid-state
