+++ acid-state-tests/acid-state-tests.cabal
@@ -1,0 +1,32 @@
+Name:                acid-state-tests
+Version:             0.15.0
+Synopsis:            Add ACID guarantees to any serializable Haskell data structure.
+Description:         Use regular Haskell data structures as your database and get stronger ACID guarantees than most RDBMS offer.
+Homepage:            https://github.com/acid-state/acid-state
+License:             PublicDomain
+Author:              David Himmelstrup
+Maintainer:          Lemmih <lemmih@gmail.com>
+-- Copyright:
+Category:            Database
+Build-type:          Simple
+Cabal-version:       >=1.10
+tested-with:         GHC == 7.6.3, GHC == 7.8.4, GHC == 7.10.3, GHC == 8.0.2, GHC == 8.2.2, GHC == 8.4.3
+
+Source-repository head
+  type:          git
+  location:      https://github.com/acid-state/acid-state
+
+Library
+  Exposed-Modules:     Data.Acid.Test.StateMachine
+
+  Build-depends:       acid-state-internal == 0.15.0,
+                       base >= 4.5.1.0 && < 5,
+                       deepseq >= 1.4.0.0,
+                       directory,
+                       hedgehog,
+                       mtl
+
+  Hs-Source-Dirs:      src/
+
+  default-language:    Haskell2010
+  GHC-Options:         -Wall
